backend/ai-service/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
albumentations==1.4.2
annotated-types==0.7.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2

chardet==5.2.0
click==8.2.1
colorlog==6.9.0
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
dataclasses-json==0.6.7
deepface==0.0.94
deepface-antispoofing==1.1.1
distro==1.9.0
dnspython==2.7.0
einops==0.2.0
email_validator==2.2.0
et_xmlfile==2.0.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fire==0.7.0
Flask==2.0.2
flask-cors==6.0.1
Flask-Session==0.4.0
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
gast==0.6.0
gdown==4.4.0

GPUtil==1.4.0
greenlet==3.2.3
grpcio
grpcio-status==1.62.3
gunicorn==23.0.0
h5py==3.14.0
hf-xet==1.1.5
httplib2==0.22.0
httptools==0.6.4
httpx-sse==0.4.1
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
keras
kiwisolver==1.4.8

lazy_loader==0.4
libclang==18.1.1
lmdb==1.7.3
lxml==6.0.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.21
mkl-service
ml-dtypes==0.3.2
more-itertools==10.7.0
msgspec==0.19.0
mtcnn==1.0.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
numpy==1.26.4
oauthlib==3.3.1
openai==1.98.0
opencv-contrib-python==4.10.0.84
opencv-python==4.8.0.76
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
optree==0.17.0
orjson==3.11.1
packaging==25.0
paddlex==3.1.3
pandas==2.3.1
pillow==10.2.0
prefetch_generator==1.0.1
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
regex==2025.7.34
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retina-face==0.0.17
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.34.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio
sounddevice==0.5.2
soupsieve==2.7
SQLAlchemy==2.0.42
starlette==0.47.2
sympy==1.13.1
tenacity==9.1.2
tensorboard
tensorboard-data-server
tensorflow==2.16.1
tensorflow-estimator
tf-keras
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.9.0
tokenizers==0.21.4
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions
tzdata==2025.2
ujson==5.10.0
ultralytics==8.3.173
ultralytics-thop==2.0.14
uritemplate==4.2.0
urllib3
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==2.0.2
wrapt==1.14.1
yarl==1.20.1
zstandard==0.23.0